+++ mod.ts
@@ -3,7 +3,13 @@ export * from "@dklab/oak-routing-ctrl";
 export * from "./src/minifyCss.ts";
 export * from "./src/ssr.ts";
 export * from "./src/types.ts";
-export * from "./src/useFullSoak.ts";
+export {
+  _unstable_useCloudflareWorkersMode,
+  type FullSoakMiddleware,
+  type OakRoutingControllerClass,
+  useFullSoak,
+  type UseFullSoakOptions,
+} from "./src/useFullSoak.ts";
 export { getOrigin } from "./src/metastore.ts";
 export { setupDefaultFullsoakLogger } from "./src/utils.ts";
 export { HtmlShell } from "./src/HtmlShell.tsx";